refactor: 패키지 분리 및 리팩터링
